# Aithon-Shield-Complete/00_Full_Source_Code/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aithon Shield API (subset)",
    "version": "1.0.0",
    "description": "REST API for scan workflows. Authenticate with session cookie `sessionId` (browser) or `Authorization: Bearer <aithon_…>` / `X-API-Key: <aithon_…>` for agents and MCP. API keys carry scopes: **read** (GET/HEAD), **write** (mutations), **admin** (full access). Create keys in Settings → API keys. Rate limits (skipped when `AITHON_DEMO_MODE` + dev): global per-IP on `/api` (default 120/min), stricter on login/signup (default 15/15 min), per-key hash (default 60/min). Responses may include `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`, `Retry-After` (429). MCP stdio: `npm run mcp` with env `AITHON_API_URL` and `AITHON_API_KEY`. Optional browser CORS for `/api`: set `AGENT_CORS_ORIGINS` to a comma-separated allowlist."
  },
  "servers": [{ "url": "/", "description": "Same origin as the web app" }],
  "components": {
    "securitySchemes": {
      "bearerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from Settings (prefix aithon_). Scopes: read (GET/HEAD), write (POST/PATCH/DELETE/…), admin (all)."
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Same value as Bearer token"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId",
        "description": "Browser session after login"
      }
    }
  },
  "security": [{ "bearerApiKey": [] }, { "apiKeyHeader": [] }, { "sessionCookie": [] }],
  "paths": {
...
    "/api/git-connections": {
      "get": {
        "summary": "List linked GitHub/GitLab accounts (no secrets; session required)",
        "security": [{ "sessionCookie": [] }]
      }
    },
    "/api/git-connections/{id}": {
      "delete": {
        "summary": "Disconnect a Git integration (session required)",
        "security": [{ "sessionCookie": [] }]
      }
    },
    "/api/remediation-jobs": {
      "get": {
        "summary": "List remediation jobs for a scan (session required)",
        "description": "Query: scanType (mvp|mobile|web|pipeline|container|network|linter), scanId.",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "name": "scanType", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "scanId", "in": "query", "required": true, "schema": { "type": "string" } }
        ]
      },
      "post": {
        "summary": "Create remediation job (opens PR/MR asynchronously; session required)",
        "description": "Requires linked Git account for provider. Demo mode: 403.",
        "security": [{ "sessionCookie": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["scanType", "scanId", "provider", "repoFullName"],
                "properties": {
                  "scanType": { "type": "string", "enum": ["mvp", "mobile", "web", "pipeline", "container", "network", "linter"] },
                  "scanId": { "type": "string" },
                  "provider": { "type": "string", "enum": ["github", "gitlab"] },
                  "repoFullName": { "type": "string", "description": "GitHub: owner/repo. GitLab: path_with_namespace or numeric project id as string" },
                  "findingIds": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/api/remediation-jobs/{id}": {
      "get": {
        "summary": "Get one remediation job (session required)",
        "security": [{ "sessionCookie": [] }]
      }
    },
...
  }
}